refactor pmodels
